+++ include/swift/Option/FrontendOptions.td
@@ -256,6 +256,9 @@ def disable_mandatory_semantic_arc_opts : Flag<["-"], "disable-mandatory-semanti
 def assume_parsing_unqualified_ownership_sil : Flag<["-"], "assume-parsing-unqualified-ownership-sil">,
   HelpText<"Assume unqualified SIL ownership when parsing SIL">;
 
+def suppress_static_exclusivity_swap : Flag<["-"], "suppress-static-exclusivity-swap">,
+  HelpText<"Suppress static violations of exclusive access with swap()">;
+
 def enable_sil_opaque_values : Flag<["-"], "enable-sil-opaque-values">,
   HelpText<"Enable SIL Opaque Values">;
